Modified coach stats in floating point representation

See merge request courses/2020-fall/csci-5308/group6/csci5308!36

+++ src/main/java/com/IceHockeyLeague/LeagueManager/Coach/ICoachStats.java
@@ -1,16 +1,15 @@
 package com.IceHockeyLeague.LeagueManager.Coach;
 
 public interface ICoachStats {
-    double getSkating();
-    void setSkating(double value);
+    float getSkating();
+    void setSkating(float value);
 
-    double getShooting();
-    void setShooting(double value);
+    float getShooting();
+    void setShooting(float value);
 
-    double getChecking();
-    void setChecking(double value);
+    float getChecking();
+    void setChecking(float value);
 
-    double getSaving();
-    void setSaving(double value);
-
+    float getSaving();
+    void setSaving(float value);
 }
